backend/requirements.txt
Flask==3.0.3
firebase-admin==6.7.0
python-dotenv==1.0.1
pika==1.3.2